requirements.txt
torch>=2.3.0
transformers>=4.41.0
datasets>=2.19.0
tokenizers>=0.19.0
accelerate>=0.30.0
einops>=0.7.0
pyyaml>=6.0.1
